refactor: isolate llmfs provider catalog
